requirements.txt
aiohttp==3.9.1
ask-sdk-core==1.19.0
ask-sdk-model==1.82.0
boto3==1.34.0
botocore==1.34.162
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
requests==2.31.0
pandas>=1.3.0
gdown>=4.4.0
opensearch-py==2.5.0
requests-aws4auth==1.2.3
